move and fix stuff
