--- psu-block/index.html
@@ -1,29 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, viewport-fit=cover" />
-  <meta name="Description" content="Put your description here.">
-  <base href="/">
-
-  <style>
-    html,
-    body {
-      margin: 0;
-      padding: 0;
-      font-family: sans-serif;
-      background-color: #ededed;
-    }
-  </style>
-  <title>psu-block</title>
-</head>
-
-<body>
-  <psu-block boxText1="8,000+ Penn State scholarships" boxText2=" " boxText3=" " stateOne></psu-block>
-  <psu-block boxText1=" " boxText2="58%" boxText3="Of students are Pennsylvania residents" stateTwo></psu-block>
-
-  <script type="module" src="./src/psu-block.js"></script>
-</body>
-
-</html>
